+++ Assets/Scripts/App.cs
@@ -1,3 +1,4 @@
+using System.Web;
 using DT.UniStart;
 using UnityEngine;
 
@@ -15,5 +16,13 @@ public class App : Entry {
     eb.AddListener("game.over", (GameOverEvent _) => {
       model.state.Value = GameState.OVER;
     });
+
+    // read server url from query string
+    if (Application.platform == RuntimePlatform.WebGLPlayer) {
+      // use next update to wait for listener to be registered
+      this.onNextUpdate.AddListener(() => {
+        eb.Invoke("set.input.serverUrl", HttpUtility.ParseQueryString(Application.absoluteURL.Split('?')[1]).Get("serverUrl"));
+      });
+    }
   }
 }