+++ Assets/Scripts/Controllers/Welcome/ServerUrlInput.cs
@@ -4,6 +4,7 @@ using TMPro;
 public class ServerUrlInput : CBC {
   void Start() {
     var config = this.Get<Config>();
+    var eb = this.Get<IEventBus>();
     var input = this.GetComponent<TMP_InputField>();
 
     // set initial value
@@ -11,5 +12,10 @@ public class ServerUrlInput : CBC {
 
     // register listener
     input.onEndEdit.AddListener((serverUrl) => config.serverUrl = serverUrl);
+
+    // listen event bus
+    eb.AddListener("set.input.serverUrl", (string serverUrl) => {
+      input.text = serverUrl;
+    });
   }
 }